Reverse unknown object / missing privilege error handling
So far we've validated if a user is allowed to see a "object unknown"
error and if not we raised a "missing privilege" error instead.

In some cases this caused information leaks. For example:

    select * from does_exist;
    MissingPrivilegeException: Missing 'DQL' privilege for user 'arthur']

    select * from does_NOT_exist;
    MissingPrivilegeException: Missing privilege for user 'arthur'

Due to the slightly different error message a user could sniff for
existing objects.

To resolve this issue this changes the logic to prefer the "object
unknown" error over a "missing privilege" error.
A user now requires "object visibility" (=any of DQL, DDL, DML) in order
to see a more detailed missing privilege error.
In any other case we throw an unknown object error, even if it does
exist.